Moved view logic to own library
